+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Additions
 Improvements
 - **Include the SpriteBatchContext in OnDrawn, OnElementDrawn and OnSelectedElementDrawn**
 - Allow scrolling panels to set height based on children by setting TreatSizeAsMaximum
+- Track element area update recursion count in UiMetrics
 
 Fixes
 - Fixed hidden scroll bars inhibiting scrolling on their parent panel